pyproject.toml
[project]
name = "2nd-innings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.2",
]
